requirements.txt
asgiref
click
cycler
Cython
Django
fonttools
gevent
greenlet
image
imutils
kiwisolver
line_profiler
matplotlib
memory-profiler
mss
numpy
obs-websocket-py
opencv-contrib-python
opencv-python
packaging
pillow
psutil
pyparsing
pyperclip
PyRect
PyScreeze
python-dateutil
python3-xlib
pytweening
six
sqlparse
websocket
websocket-client
websockets
zope.event
zope.interface
